+++ packages/bootstrap-4/src/AddButton/AddButton.tsx
@@ -4,6 +4,7 @@ import {
   IconButtonProps,
   RJSFSchema,
   StrictRJSFSchema,
+  TranslatableString,
 } from "@rjsf/utils";
 import Button from "react-bootstrap/Button";
 import { BsPlus } from "@react-icons/all-files/bs/BsPlus";
@@ -13,12 +14,13 @@ export default function AddButton<
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any
 >({ uiSchema, registry, ...props }: IconButtonProps<T, S, F>) {
+  const { translateString } = registry;
   return (
     <Button
       {...props}
       style={{ width: "100%" }}
       className={`ml-1 ${props.className}`}
-      title="Add Item"
+      title={translateString(TranslatableString.AddItemButton)}
     >
       <BsPlus />
     </Button>