+++ packages/core/src/components/templates/ErrorList.tsx
@@ -5,6 +5,7 @@ import {
   RJSFValidationError,
   RJSFSchema,
   StrictRJSFSchema,
+  TranslatableString,
 } from "@rjsf/utils";
 
 /** The `ErrorList` component is the template that renders the all the errors associated with the fields in the `Form`
@@ -15,11 +16,14 @@ export default function ErrorList<
   T = any,
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any
->({ errors }: ErrorListProps<T, S, F>) {
+>({ errors, registry }: ErrorListProps<T, S, F>) {
+  const { translateString } = registry;
   return (
     <div className="panel panel-danger errors">
       <div className="panel-heading">
-        <h3 className="panel-title">Errors</h3>
+        <h3 className="panel-title">
+          {translateString(TranslatableString.ErrorsLabel)}
+        </h3>
       </div>
       <ul className="list-group">
         {errors.map((error: RJSFValidationError, i: number) => {